% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGGAdd.R
\name{getKEGGGeneMotif}
\alias{getKEGGGeneMotif}
\title{Get motif information from KEGG}
\usage{
getKEGGGeneMotif(geneID, hasAddInfo = FALSE)
}
\arguments{
\item{geneID}{A single KEGG gene ID}

\item{hasAddInfo}{A logic element whether to show the additional information. The default value is "FALSE".}
}
\value{
A list if "hasAddInfo" is TRUE, else a matrix. A special case retuns "NULL" instead of a matrix (other information is not effected), if there is no motif information.
}
\description{
KEGG Database Additional API - Get the gene motif table
}
\details{
Get the gene motif tables and additional information.
}
\examples{
getKEGGGeneMotif('brp:103873230')

# no motif information
getKEGGGeneMotif('hsa:4558', hasAddInfo = TRUE)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

