% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getKEGGKO}
\alias{getKEGGKO}
\title{Get KEGG orthology.}
\usage{
getKEGGKO(KOID)
}
\arguments{
\item{KOID}{The KEGG orthology ID.}
}
\value{
A character vector of KEGG gene IDs
}
\description{
KEGG Database API - Get the KEGG orthology list.
}
\details{
Get the KEGG orthology list by a given KEGG KO ID.
}
\examples{
KOMat <- getKEGGKO('K02110')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.kegg.jp/kegg/rest/}
}

