% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getKEGGPathAnno}
\alias{getKEGGPathAnno}
\title{List pathway of a given species ID}
\usage{
getKEGGPathAnno(KEGGspec)
}
\arguments{
\item{KEGGspec}{The KEGG species ID. Only one species ID once input.}
}
\value{
A matrix of pathway ID and annotation.
}
\description{
KEGG Database API - Get the whole pathway ID from KEGG database.
}
\details{
Get the pathway ID and annoation of a given KEGG species ID.
}
\examples{
hasPath <- getKEGGPathAnno('hsa')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.kegg.jp/kegg/rest/}
}

