% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getKEGGPathGenes}
\alias{getKEGGPathGenes}
\title{List pathways and genes of a given KEGG species ID}
\usage{
getKEGGPathGenes(specID)
}
\arguments{
\item{specID}{KEGSS species org code or T number , for example "hsa" or "T01001".}
}
\value{
A List named with KEGG pathway IDs, and each element of the list contains the KEGG gene IDs.
}
\description{
KEGG Database API - Get the pathway and genes.
}
\details{
Get the pathway and genes according to KEGG species ID.
}
\examples{
hasPathGenes <- getKEGGPathGenes('hsa')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.kegg.jp/kegg/rest/}
}

