% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGGAdd.R
\name{getKEGGSpeInfo}
\alias{getKEGGSpeInfo}
\title{Get KEGG organism basic information}
\usage{
getKEGGSpeInfo(specID)
}
\arguments{
\item{specID}{KEGSS species org code or T number , for example "hsa" or "T01001".}
}
\value{
A two-column matrix contains basic information.
}
\description{
KEGG Database Additional API - Get KEGG organism basic information.
}
\details{
The KEGG organism basic information is retrieved from the webpage.
}
\examples{
hasInfo <- getKEGGSpeInfo('hsa')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.genome.jp/kegg-bin/show_organism?org=T01001}
}

