% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGGAdd.R
\name{getKEGGTIDGeneSeq}
\alias{getKEGGTIDGeneSeq}
\title{KEGG Database Additional API - Get mutiple nucleotide acid and amino acid sequences according to the T numbers}
\usage{
getKEGGTIDGeneSeq(TIDs, seqType = "aaseq", n = 4)
}
\arguments{
\item{TIDs}{A vector of T number IDs. for the protein or gene.}

\item{seqType}{Choose nucleotide acid (ntseq) or amino acid (aaseq) seqences, and the default is amino acid sequences.}

\item{n}{The number of CPUs or processors, and the default value is 4.}
}
\value{
A BStringSet
}
\description{
Get mutiple protein and gene sequences from KEGG T number in fasta format. As there is no direct API for retrieving the sequence from T number, for example "T10017:100009". The fasta sequence is extract from a webpage like "http://www.genome.jp/dbget-bin/www_bget?-f+-n+a+t10017:100009".
}
\examples{
tNumMultiSeqs <- getKEGGTIDGeneSeq(c('T10017:100009', 'T10017:100036', 'T10017:100044'), n = 2)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.genome.jp/dbget-bin/www_bget?-f+-n+a+t10017:100009}

\url{http://www.genome.jp/dbget-bin/www_bget?-f+-n+n+t10017:100009}
}
\seealso{
getKEGGGeneSeq
}

