% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getProID}
\alias{getProID}
\title{KEGG Database API - Get the whole KEGG IDs from one species.}
\usage{
getProID(KEGGspec)
}
\value{
A matrix of KEGG IDs and annotation
}
\description{
KEGG Database API - Get the whole KEGG IDs from one species.
}
\examples{
## KEGG org cord
getProID('eco')

## KEGG T number
getProID('T00007')

## KEGG T number with empty elements
getProID('T10004')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.kegg.jp/kegg/rest/}
}

