% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hello.R
\name{hello}
\alias{hello}
\title{Hello, Von Weber!}
\usage{
hello()
}
\description{
This is an example function named 'hello',
which prints 'Hello, Von Weber!'.
}
\details{
You can learn more about package authoring with RStudio at:

\url{http://r-pkgs.had.co.nz/}
}
\examples{
hello()
}
