% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kegg_name2id.R
\name{kegg_name2id}
\alias{kegg_name2id}
\title{Convert (standard) names to kegg id}
\usage{
kegg_name2id(names = "", database = "ko", ignore.case = FALSE)
}
\arguments{
\item{names}{list of names to match.}

\item{database}{string. Database to search in.}

\item{ignore.case}{logical. Case sensitivity or not.}
}
\value{
dataframe. \code{found} represents different status. \code{0}: doesn't
find something matched. \code{1}: found only 1 entry match. \code{2}: found more
than 1 entries matched, and we only take the first matched id.
}
\description{
As the stardand names are not unique in a sense. Here we refer to the characters
that wo can match in kegg database. And if \code{database} is specified to concrete
sub-database, conversion is faster.
}
\examples{
# "Rb1" and "pRb" are the same gene of mouse(mmu)
kegg_name2id(c("aaaa","Rb1","pRb"), "mmu")
}
