% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggColor.R
\name{keggColor}
\alias{keggColor}
\title{Color Pathway}
\usage{
keggColor(map, mapping_list = NULL, mode = "color", log = 0,
          numericalType = "mm", minColor = "#ffff00", maxColor = "#ff0000",
          negativeColor = "#00ff00", zeroColor = "#ffff00", positiveColor = "#ff0000")

keggColor(map, filename = NULL, mode = "color", log = 0,
          numericalType = "mm", minColor = "#ffff00", maxColor = "#ff0000",
          negativeColor = "#00ff00", zeroColor = "#ffff00", positiveColor = "#ff0000")
}
\arguments{
\item{map}{string. Pathway ID.}

\item{mapping_list}{string. The coded entries and colors/numbers.}

\item{filename}{string. File that contains entries and colors/numbers, same effect as mapping_list.}

\item{mode}{"color" or "number". Use "Color specification" or "Numerical value".}

\item{log}{1 or 0. Convert to log scale or not.}

\item{numericalType}{"mm" or "nzp". Gradation with minimum-maximum or negative-zero-positive.}

\item{minColor, maxColor}{Give the colors at minimum and maixmum.}

\item{negativeColor, zeroColor, positiveColor}{Give the colors at negative, zero and positive.}

\item{path}{string. Path to save images.}
}
\description{
Color Pathway is an advanced version of the KEGG pathway mapping tool,
where given objects (genes, proteins, compounds, glycans, reactions, drugs, etc.)
are searched against KEGG pathway maps and found objects are marked in any background
and foreground colors (bgcolor and fgcolor).
}
\details{
For each entry, the format is \code{/keggid\%09bgcolor,fgcolor}
or \code{/keggid\%09number} in mapping_list, the format is \code{keggid bgcolor,fgcolor}
or \code{keggid number} in each line in the file.

When mode = "number" and numericalType = "mm"(Gradation with minimum-maximum),
we just need minColor and maxColor.
When mode = "number" and numericalType = "nzp"(Gradation with negative-zero-positive),
we just need negativeColor, zeroColor and positiveColor.
When mode = "color", we need none of them.
}
\examples{
# "\%09" represents TAB in ASCII code
# Use "\%23" in ASCII code instead of "#" for color specification.
# color mode
keggColor(map = "map00400",
          mapping_list = "/1.14.16.1\%09,blue/C00079\%09,red/C00166\%09\%23005050")

# numer mode
keggColor(map = "map00400",
          mapping_list = "/1.14.16.1\%091/C00079\%092/C00166\%093",
          mode = "number")

# give filename, color mode
keggColor(map = "hsa05200",
     filename = "data/hsa_CML.txt")

# give filename, number mode
keggColor(map = "hsa05200",
          filename = "data/hsa_CML-COSMIC.txt",
          mode = "number")


}
