% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggConv.R
\name{keggConv}
\alias{keggConv}
\title{Convert IDs}
\usage{
keggConv(target_db, source_db)
keggConv(target_db, dbentries)
}
\arguments{
\item{target_db}{string. The database convert to.}

\item{source-db}{string. The database convert from.}

\item{dbentries}{string/char array. The IDs of entry we want to convert.}
}
\description{
Convert KEGG identifiers to/from outside identifiers.
}
\details{
When converting from database to database,
\code{(target_db, source_db) = (kegg_db, outside_db) | (outside_db, kegg_db)}.
For gene identifiers: \code{kegg_db = org, org = KEGG organism code or T number,
outside_db = ncbi-proteinid | ncbi-geneid | uniprot}.

For chemical substance identifiers: \code{kegg_db = drug | compound | glycan
outside_db = pubchem | chebi}.

When converting entries to target_db, For gene identifiers:
\code{<dbentries> = database entries involving the following <database>,
<database> = <org> | genes | ncbi-proteinid | ncbi-geneid | uniprot,
<org> = KEGG organism code or T number}

For chemical substance identifiers:
\code{<dbentries> = database entries involving the following <database>
<database> = drug | compound | glycan | pubchem | chebi}
}
\note{
The database name "genes" may be used only in the second form
(convert entries to target_db).
}
\examples{
# conversion from NCBI GeneID to KEGG ID for E. coli genes
gene_E.coli = keggConv('eco', 'ncbi-geneid')
# opposite direction
gene_E.coli = keggConv('ncbi-geneid', 'eco')

# conversion from KEGG ID to NCBI ProteinID
keggConv('ncbi-proteinid', 'hsa:10458+ece:Z5100')

# conversion from NCBI GeneID to KEGG ID when the organism code is not known
keggConv('genes', 'ncbi-geneid:19831932')
# same as above
keggConv('lpn', 'ncbi-geneid:19831932')
}
