% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggFind.R
\name{keggFind}
\alias{keggFind}
\title{Search Data in KEGG Database}
\usage{
keggFind()
}
\arguments{
\item{keywords}{string. Query keywords or other query data with given database.}
}
\description{
Finds entries with matching query keywords
or other query data in a given database
}
\details{
URL Form 1:

http://rest.kegg.jp/find/<database>/<query>

<database> = pathway | module | ko | genome | <org> | compound | glycan |
reaction | rclass | enzyme | disease | drug | dgroup | environ |
genes | ligand

<org> = KEGG organism code or T number

URL Form 2:

http://rest.kegg.jp/find/<database>/<query>/<option>

<database> = compound | drug

<option> = formula | exact_mass | mol_weight
}
