% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggGetImage.R
\name{keggGetImage}
\alias{keggGetImage}
\title{Get Images from KEGG}
\usage{
keggGetImage(keywords, path = getwd(), display = FALSE)
}
\arguments{
\item{keywords}{string, or character array. Corresponding IDs.}

\item{path}{string. The path to save images.}

\item{dispaly}{logical. Display images in R window or not.}
}
\description{
Get Images of Pathways, Compounds, etc, from KEGG
}
\details{
Using '+' to add multiple IDs in keywords string.

Images of pathway won't dispaly in R window, causing it's usually too large.
They will be saved in path \code{/image} by default. Of course, all images from
\code{keggGetImage()} will be saved.
}
\examples{
# retrieves the png image file of a pathway map
keggGetImage('hsa05130')

# retrieves the gif image file of a compound
keggGetImage('C00002')

# retrieves image files of multiple pathway maps or compoumnds.
keggGetImage(c('hsa05130', 'C00002'))
# same as above
keggGetImage('hsa05130+C00002')

}
