% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggInfo.R
\name{keggInfo}
\alias{keggInfo}
\title{Get KEGG Database Infomation}
\usage{
keggInfo()
}
\value{
A dataframe of how many entries in each sub-database.
}
\description{
Displays the current statistics of a given database.
}
\examples{
keggInfo()
}
