% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggLink.R
\name{keggLink}
\alias{keggLink}
\title{Linked Entries}
\usage{
keggLink("target_db", "source_db")
keggLink("target_db", "dbentries")
}
\arguments{
\item{target_db, source_db}{string. database I}

\item{dbentries}{string. KEGG database entries involving the database II}
}
\description{
Find related entries by using database cross-references.
}
\details{
database I = pathway | brite | module | ko | genome | <org> |
compound | glycan | reaction | rclass | enzyme | disease | drug |
dgroup | environ

database II = pathway | brite | module | ko | genome | <org> |
compound | glycan | reaction | rclass | enzyme | disease |
drug | dgroup | environ | genes
}
\note{
The database name "genes" may be used only in the second form.
}
\examples{
# KEGG pathways linked from each of the human genes
keggLink("pathway", "hsa")

# human genes linked from each of the KEGG pathways
keggLink("hsa", "pathway")

# KEGG pathways linked from a human gene and an E. coli O157 gene
keggLink("pathway", "hsa:10458+ece:Z5100")

# List of genes with the KO assignment of K00500
keggLink("genes", "K00500")

# List of human genes in pathway hsa00010
keggLink("genes", "hsa00010")
# or
keggLink("hsa", "hsa00010")

# List of KO entries in pathway map00010 or ko00010
keggLink("ko", "map00010")
keggLink("ko", "ko00010")

# List of reaction entries in pathway map00010 or rn00010
keggLink("rn", "map00010")
keggLink("rn", "rn00010")

# List of EC number entries in pathway map00010 or ec00010
keggLink("ec", "map00010")
keggLink("ec", "ec00010")

# List of compound entries in pathway map00010
keggLink("cpd", "map00010")

}
