% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggList.R
\name{keggList}
\alias{keggList}
\title{Get KEGG Database or Entry List}
\usage{
keggList(x)
}
\arguments{
\item{x}{string. Database(s) or entry[ies], using '+' to join.}
}
\value{
dataframe. The table of particular contents.
}
\description{
Returns a list of entry identifiers and associated definition
for a given database or a given set of database entries.
}
\examples{
# returns the list of reference pathways
keggList('pathway')

# returns the list of human pathways
keggList('pathway/hsa')

# returns the list of a human gene and an E.coli O157 gene
keggList('hsa:10458+ece:Z5100')

}
