% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGGAdd.R
\name{getKEGGMotifList}
\alias{getKEGGMotifList}
\alias{getKEGGMotifList2}
\title{Get motif list from KEGG.}
\usage{
getKEGGMotifList(motifName)

getKEGGMotifList2(motifName)
}
\arguments{
\item{motifName}{A single KEGG motif ID}
}
\value{
A matrix of KEGG genes and description

A vector of protein names including UniProt and SWISS-PROT. Not all these "protID" have KEGG IDs. KEGG uses UniProt and SWISS-PROT for gene UniProt annotation.
}
\description{
KEGG Database Additional API - Get genes processing certain motif
}
\details{
Get all the gene having certain motif.
}
\examples{
\dontrun{
getKEGGMotifList('pf:DUF3675')}
pfUniprot <- getKEGGMotifList2('pf:DUF3675')
\dontrun{
# convert uniprot ID to KEGG IDs
pfUniprot <- paste0('uniprot:', pfUniprot)
pfGeneID <- convKEGG('genes', pfUniprot, convertType = 'identity')
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
\seealso{
convKEGG
}

