% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGGAdd.R
\name{singleTIDSeq}
\alias{singleTIDSeq}
\title{Get protein and gene sequences from T numbers}
\usage{
singleTIDSeq(TID, seqType = "aaseq")
}
\arguments{
\item{TID}{The T number ID for the protein or gene.}

\item{seqType}{Choose nucleotide acid ('ntseq') or amino acid ('aaseq') seqences, and the default is amino acid sequences.}
}
\value{
A BStringSet
}
\description{
KEGG Database Additional API - Get nucleotide acid and amino acid sequences according to the T numbers
}
\details{
Get protein and gene sequences from KEGG T number in fasta format. As there is no direct API for retrieving the sequence from T number, for example "T10017:100009". The fasta sequence is extract from a webpage like "http://www.genome.jp/dbget-bin/www_bget?-f+-n+a+t10017:100009". The function singleTIDSeq() get a sequence one time, and the function getKEGGTIDGeneSeq() provides a parallel way to download sequences.
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

