% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/misc.R
\name{webTable}
\alias{webTable}
\title{Get R matrix from weblink}
\usage{
webTable(url, ncol)
}
\arguments{
\item{url}{The weblink.}

\item{ncol}{The column number of the matrix.}
}
\value{
A R matrix
}
\description{
Get a R matrix object if the weblink returned as a matrix.
}
\details{
If the web return a matrix, use this function to extract it as a R matrix object. An empty element is used to represent a "NA" in KEGG. So the empty web element is set to "NA" according the "ncol" parameter.
}
\examples{
pathUrl <- 'http://rest.kegg.jp/list/pathway'
keggPath <- webTable(pathUrl, ncol = 2)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

