% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{EnforceGetURL}
\alias{EnforceGetURL}
\title{Enforce getting url}
\usage{
EnforceGetURL(tryUrl, FUN)
}
\arguments{
\item{tryUrl}{url}

\item{FUN}{Functions that can readin url, for example the "getURL()" function in the "RCurl" package.}
}
\value{
retured value from getURL()
}
\description{
Enforce to get URL
}
\details{
If any errors occur when getting the url, this functions holds the error and try to get url again. An infinit loop may happen if the input url could not be resoved.
}
\examples{
\dontrun{
## It will cause infinit loop
require(RCurl)
testUrl <- 'http://www.test1111111111.com/'
EnforceGetURL(testUrl, FUN = getURL)
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

