% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractComponent.R
\name{extractComponent}
\alias{extractComponent}
\title{extract all components from a kegg pathway}
\usage{
extractComponent(url)
}
\arguments{
\item{url}{string. String of pathway ID or detailed page.}
}
\description{
extract all components from a kegg pathway
}
\examples{
c1 = extractComponent("mmu04933")
# equivalent to
url = "http://www.kegg.jp/kegg-bin/highlight_pathway?map=mmu04933"
c2 = extractComponent(url)


}
