% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{getKEGGGeneInfo}
\alias{doTenInfo}
\alias{getKEGGGeneInfo}
\title{Get gene informatino}
\usage{
getKEGGGeneInfo(KEGGGeneIDs, n = 1)

doTenInfo(tenWebInfo)
}
\arguments{
\item{KEGGGeneIDs}{A vector of KEGG IDs. Seqences from different species could be combined together.}

\item{n}{The number of CPUs or processors, and the default value is 1.}

\item{tenWebInfo}{return results from function getURL().}
}
\value{
A string vectors.

a list each contains the annotation items (for the case without annotation returns NULL).
}
\description{
KEGG Database API - Get gene information
}
\details{
Get gene information. This function supports multiple querys.
}
\examples{
genes <- c('eco:b4600', 'ece:Z5100', 'eco:b3160', 'dra:DR_0001', 'dra:DR_A0001', 'dra:DR_B0001')
genesInfo <- getKEGGGeneInfo(genes, n = 2)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

