% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{getKEGGGeneInfo}
\alias{ExtractLocus}
\alias{ExtractMotif}
\alias{doTenInfo}
\alias{getKEGGGeneInfo}
\title{Get gene information}
\usage{
getKEGGGeneInfo(KEGGGeneIDs, enforceURL = FALSE, n = 1)

doTenInfo(tenWebInfo)

ExtractLocus(locusStr)

ExtractMotif(motifStr)
}
\arguments{
\item{KEGGGeneIDs}{A vector of KEGG IDs. Seqences from different species could be combined together.}

\item{enforceURL}{whether to enfoce get get url until the results returned. The default value is "FALSE".}

\item{n}{The number of CPUs or processors, and the default value is 1.}

\item{tenWebInfo}{return results from function getURL().}

\item{locusStr}{named location strings whose names are gene names.}

\item{motifStr}{named motif strings whole names are gene names}
}
\value{
getKEGGGeneInfo(): A string vectors containing gene informations

ExtractLocus(): A matrix: 1st gene names, 2ed is genome,  3rd is start, 4th is end, and the 5th is the strand.

ExtractMotif(): A list: each element is a motif vector.

a list each contains the annotation items (for the case without annotation returns NULL).
}
\description{
KEGG Database API - Get gene information
}
\details{
getKEGGGeneInfo(): Get gene information. This function supports multiple querys.

ExtractLocus(): Extract prokaryotic gene location.

ExtractMotif(): Extract protein motifs.
}
\examples{
genes <- c('eco:b4600', 'ece:Z5100', 'eco:b3160', 'dra:DR_0001', 'dra:DR_A0001', 'dra:DR_B0001')
genesInfo <- getKEGGGeneInfo(genes, n = 2)
locusVec <- c('complement(join(1631002..1632285,1652755..1652838))', 'complement(4658240..4658986)',
'3303448..3304455', '1:complement(1..1182)', '2:653..1435', 'MP1:520..1296')
names(locusVec) <- c('eco:b4600', 'ece:Z5100',
'eco:b3160', 'dra:DR_0001', 'dra:DR_A0001', 'dra:DR_B0001')
ExtractLocus(locusVec)
motifVec <- c('Pfam: B_lectin S_locus_glycop',
'Pfam: VIT1 AlaDh_PNT_C DUF4168 PsbI',
'Pfam: BRF1 Coilin_N')
names(motifVec) <- c('ath:AT2G01780', 'ath:AT2G01770',
'ath:AT2G01800')
ExtractMotif(motifVec)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

