% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/geneInfo.R
\name{getKEGGGeneInfo}
\alias{ExtractProLocus}
\alias{doTenInfo}
\alias{getKEGGGeneInfo}
\title{Get gene informatino}
\usage{
getKEGGGeneInfo(KEGGGeneIDs, enforceURL = FALSE, n = 1)

doTenInfo(tenWebInfo)

ExtractProLocus(locusStr)
}
\arguments{
\item{KEGGGeneIDs}{A vector of KEGG IDs. Seqences from different species could be combined together.}

\item{enforceURL}{whether to enfoce get get url until the results returned. The default value is "FALSE".}

\item{n}{The number of CPUs or processors, and the default value is 1.}

\item{tenWebInfo}{return results from function getURL().}

\item{locusStr}{named location strings whose names are gene names.}
}
\value{
A string vectors.

a list each contains the annotation items (for the case without annotation returns NULL).

a 5-column matrix, 1st gene names, 2ed is genome,  3rd is start, 4th is end, and the 5th is the strand.
}
\description{
KEGG Database API - Get gene information
}
\details{
getKEGGGeneInfo(): Get gene information. This function supports multiple querys.

ExtractProLocus(): Extract prokaryotic gene location.
}
\examples{
genes <- c('eco:b4600', 'ece:Z5100', 'eco:b3160', 'dra:DR_0001', 'dra:DR_A0001', 'dra:DR_B0001')
genesInfo <- getKEGGGeneInfo(genes, n = 2)
locusVec <- c('complement(join(1631002..1632285,1652755..1652838))', 'complement(4658240..4658986)',
'3303448..3304455', '1:complement(1..1182)', '2:653..1435', 'MP1:520..1296')
names(locusVec) <- c('eco:b4600', 'ece:Z5100',
'eco:b3160', 'dra:DR_0001', 'dra:DR_A0001', 'dra:DR_B0001')
ExtractProLocus(locusVec)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}

Yulong Niu \email{niuylscu@gmail.com}
}
\keyword{internal}

