% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getKEGGPathGenes}
\alias{getKEGGPathGenes}
\title{List pathways and genes of a given KEGG species ID}
\usage{
getKEGGPathGenes(KEGGspec)
}
\arguments{
\item{KEGGspec}{The KEGG species ID. Only one species ID once input.}
}
\value{
A List named with KEGG pathway IDs, and each element of the list contains the KEGG gene IDs.
}
\description{
KEGG Database API - Get the pathway and genes.
}
\details{
Get the pathway and genes according to KEGG species ID.
}
\examples{
\dontrun{
getKEGGPathGenes('hsa')}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

