% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getKEGGPhylo}
\alias{getKEGGPhylo}
\title{Get species list from KEGG.}
\usage{
getKEGGPhylo(speList, speType = "KEGG", whole = FALSE)
}
\arguments{
\item{speList}{The species list that is a vector like "c('hsa', 'eco')". The input "speList" should be consistent with the parameter "speType".}

\item{speType}{It supports five types: "KEGG", "Tnum", "regexpr", and "phylo".
KEGG type is a three or four letters, for exmaple "hsa" is the KEGG ID for Homo sapiens,}

\item{whole}{Whether or not get the whole KEGG species list,
and the default value is FALSE.}
}
\value{
Matrix of species information.
}
\description{
KEGG Database API - Get species KEGG/NCBI ID.
}
\details{
Get the phylogenetic information of given species.
It supports both batch input and regular expression search.
}
\examples{
## search species list from KEGG ID
getKEGGPhylo(c('hsa', 'eco'))
## search species whose names include 'Escherichia coli'
getKEGGPhylo('Escherichia coli', speType = 'regexpr')
## search species whose class is 'Mammals'
getKEGGPhylo('Mammals', speType = 'phylo')
## get whole KEGG species information table
getKEGGPhylo(whole = TRUE)
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.kegg.jp/kegg/rest/}
}

