% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKEGGAdd.R
\name{getKEGGSpeInfo}
\alias{getKEGGSpeInfo}
\title{Get KEGG organism basic information}
\usage{
getKEGGSpeInfo(specID)
}
\arguments{
\item{specID}{KEGSS species org code or T number , for example "hsa" or "T01001".}
}
\value{
A list contains basic information.
}
\description{
KEGG Database Additional API - Get KEGG organism basic information
}
\details{
The KEGG organism basic information is retrieved from the webpage. If references are included, a four-row matrix ("Reference", "Authors", "Title", "Journal") will be returned. If chromosome are included, a three-row matrix ("Chromosome", "Sequence", and "Length") will be returned. The similar process will be applied to plasmid.
}
\examples{
hasInfo <- getKEGGSpeInfo('hsa')
draInfo <- getKEGGSpeInfo('dra')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.genome.jp/kegg-bin/show_organism?org=T01001}
}

