% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getKEGG.R
\name{getProID}
\alias{getProID}
\title{Get whole KEGG IDs and annotation}
\usage{
getProID(specID, ...)
}
\arguments{
\item{specID}{KEGSS species org code or T number , for example "hsa" or "T01001".}

\item{...}{Parameters inherited from webTable().}
}
\value{
A matrix of KEGG IDs and annotation
}
\description{
KEGG Database API - Get the whole KEGG IDs from one species.
}
\details{
Get the KEGG protein ID list and annotation.
}
\examples{
## KEGG org cord
getProID('eco')

\dontrun{
## KEGG T number
getProID('T00007')}

## KEGG T number with empty elements
getProID('T10004')
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}
\references{
\url{http://www.kegg.jp/kegg/rest/}
}

