% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getKEGGAdd.R
\name{transPhyloKEGG2NCBI}
\alias{transPhyloKEGG2NCBI}
\title{Get NCBI Taxonomy ID From KEGG ID}
\usage{
transPhyloKEGG2NCBI(KEGGID, n = 4)
}
\arguments{
\item{KEGGID}{The KEGG support multiple species ID, for example c('hsa', 'eco').}

\item{n}{The number of CPUs or processors, and the default value is 4.}
}
\value{
The corresponding NCBI Taxonomy ID in character vector.
}
\description{
KEGG Database Additional API - Get the NCBI taxonomy ID from a given KEGG ID
}
\details{
NCBI taxonomy ID is used as unique ID accoss KEGG and BioCyc databases. This functions is used to get the corresponding NCBI Taxonomy ID from KEGG.
}
\examples{
# get human and Ecoli NCBI taxonomy ID with 2 threads
transPhyloKEGG2NCBI(c('hsa', 'eco', 'ath', 'smu'), n = 2)

\dontrun{
# transfer all KEGG species ID to NCBI taxonomy ID
wKEGGSpe <- getKEGGPhylo(whole = TRUE)
wNCBISpe <- transPhyloKEGG2NCBI(wKEGGSpe[, 2])
}
}
\author{
Yulong Niu \email{niuylscu@gmail.com}
}

